+++ src/account/models/__init__.py
@@ -1,0 +1,3 @@
+from .barber import *
+from .customer import *
+from .custom_user import *
